% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_reg.R
\name{linear_reg}
\alias{linear_reg}
\title{A wrap of linear regression}
\usage{
linear_reg(formula, data)
}
\arguments{
\item{formula:}{a formula}

\item{data:}{dataset}
}
\value{
a lm object
}
\description{
This function performs linear regression (a simplified version)
}
\examples{
linear_reg(Sepal.Length~., iris)
}
